% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{bscale}
\alias{bscale}
\title{Backscale an array}
\usage{
bscale(A, cs)
}
\arguments{
\item{A}{A 3D landmark array}

\item{cs}{a vector of centroid sizes}
}
\value{
Returns a scaled 3D landmark array
}
\description{
Scale each individual of a landmark array by an arbitrary value (e.g. centroid size)
}
