% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVA-shape.R
\name{CV.3D.shape}
\alias{CV.3D.shape}
\title{CVA 3D Shape Chance}
\usage{
CV.3D.shape(CVA, variate = 1, scale = 5, links = NULL)
}
\arguments{
\item{CVA}{object of class CVA generated by Morpho::CVA}

\item{variate}{The canonical variate to be visualized (default = 1)}

\item{scale}{A scaling factor to emphasize shape change (default = 5)}

\item{links}{A wireframe for visualizing. WIP}
}
\value{
Plots the mean landmark configuration of a CVA an vectors representing the shape change along a given variate, by a given scale
}
\description{
Visualize Shape change in three dimenions of Morpho object.class 'CVA'
}
