% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{grp.apply}
\alias{grp.apply}
\title{Group Apply}
\usage{
grp.apply(d, f, grp)
}
\arguments{
\item{d}{a dataset}

\item{f}{a function}

\item{grp}{a grouping varaible}
}
\description{
Apply a function (f) to a set of data (d) by a grouping variable (grp).
}
