% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{lmk.swap}
\alias{lmk.swap}
\title{Swap Landmarks}
\usage{
lmk.swap(l1, l2)
}
\arguments{
\item{l1}{a numeric vector of landmark index (sequence number) to be changed}

\item{l2}{a numeric vector of landmark index (Sequence number) to change to}
}
\description{
Function to swap landmark configurations, or set as missing
}
\author{
Daniel Ehrlich
}
