% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_colpick}
\alias{LMK_colpick}
\title{Graphical Color Picker}
\usage{
LMK_colpick(n = 1, type = c("shades", "tints"), transparency = 1)
}
\arguments{
\item{n}{Number of colors to pick}

\item{type}{One of c("shades, "tints") to plot}

\item{transparency}{Set color transparency (alpha) from 0 (invisible) to 1 (opaque). Default is opaque}
}
\value{
Returns a matrix of n colors listed in 3 formats: Rcolor, RGB, and HSV.
}
\description{
Interactive function to choose precise color from a pallete of tints or shades
}
\details{
Select any number of colors from palette
}
