% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMK_Viz.R
\name{LMK_CVA_2D}
\alias{LMK_CVA_2D}
\title{CVA 2D Shape Change}
\usage{
LMK_CVA_2D(CVA, variate = 1, scale = 5)
}
\arguments{
\item{CVA}{object of class CVA generated by Morpho::CVA}

\item{variate}{The canonical variate to be visualized (default = 1)}

\item{scale}{A scaling factor to emphasize shape change (default = 5)}
}
\value{
Plots the mean landmark configuration of a CVA an vectors representing the shape change along a given variate, by a given scale
}
\description{
Visualize Shape change in two dimensions of Morpho object.class 'CVA'
}
