% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_get_sliders}
\alias{LMK_get_sliders}
\title{Get sliders}
\usage{
LMK_get_sliders(c1)
}
\arguments{
\item{c1}{A vector of landmark numbers}
}
\value{
Returns a table of sliders suitable for use with geomorph::gpagen(). This function assumes all points slide sequentially between the max/min landmarks in the sequence.
}
\description{
Define a sliders table (geomorph) for a contour with fixed endpoints
}
\details{
Assumes all points slide sequentially between endpoints
}
\examples{

lmknums <- 1:100

sli <- LMK_get_sliders(lmknums[80:100]) ## to treat landmarks 80 through 100 as sliders

sli

}
\author{
Daniel Ehrlich
}
