% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_iccA}
\alias{LMK_iccA}
\title{Intra class correlations and Chronbach's Alpha}
\usage{
LMK_iccA(obs1, obs2)
}
\arguments{
\item{obs1}{a 3D landmark array}

\item{obs2}{a 3D landmark array}
}
\value{
Returns a table of ICC coefficients, associated p-values, and Chronbach's Alpha
}
\description{
Compute ICC and CA for each dimension (X,Y,Z) of arrays
}
