% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_limset}
\alias{LMK_limset}
\title{limset}
\usage{
LMK_limset(x, factor = 1.2)
}
\arguments{
\item{x}{A vector to be plotted}

\item{factor}{to expand the data (default=1.2)}
}
\description{
Function to easily set plot buffer
}
\details{
Takes a vector of data, calculates the range of data and scales it by a factor. Returns new range of length=2
}
\examples{

vec <- 1:10
extend_vec <- LMK_limset(vec)
extend_vec

}
