% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMK_Viz.R
\name{LMK_PCA_plot}
\alias{LMK_PCA_plot}
\title{Shapespace PCA}
\usage{
LMK_PCA_plot(A, xPC = 1, yPC = 2, zPC = NULL, grp)
}
\arguments{
\item{A}{A 3d landmark array}

\item{xPC}{an integer indicating which PC to plot, default = 1}

\item{yPC}{an integer indicating which PC to plot, default = 2}

\item{zPC}{an integer indicating which PC to plot, default = NULL (2D)}

\item{grp}{a grouping factor}
}
\description{
Plot 2D or 3D scatter plot of tangent shape-space for a GPA object (of gpagen())
}
