% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-outlier.R
\name{LMK_plotoutliers}
\alias{LMK_plotoutliers}
\title{Plot Outliers}
\usage{
LMK_plotoutliers(a, gpa = TRUE, ...)
}
\arguments{
\item{a}{Raw 3D coordinate data, procrustes aligned coordinate data, or an object of class == gpagen}

\item{gpa}{Logical, should a GPA be calculated. If TRUE, a is assumed to be raw coordinate data and GPA is implemented via geomorph::gpagen. If FALSE, no imposition will be done.}

\item{...}{Additional parameters to pass to gpagen}
}
\description{
Function to identify outliers visually
}
