% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMK_screen.R
\name{LMK_screen}
\alias{LMK_screen}
\title{Screen Data}
\usage{
LMK_screen(A, ref = NULL, gpa = TRUE)
}
\arguments{
\item{A}{a 3D array of landmark data in the form p x k x n}

\item{ref}{a known standard to compare to (if null (default), function will calculate group mean)}

\item{gpa}{logical, should a GPA be implemented first. If TRUE, data are assumed to be raw landmark coordinates. If GPA = FALSE, data are assumed to be already aligned procrustes coordinates}
}
\value{
Returns a vector of individuals
}
\description{
Interactive function to compare individuals in an array to grandmean (or other refrence) to spot misplaced landmarks/other outliers.
}
\details{
Visually inspect a 3D landmark array for outliers/other data inconsistencies. Create a list of potential individuals to inspect further.
}
