% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_swap}
\alias{LMK_swap}
\title{Swap Landmarks}
\usage{
LMK_swap(a, l1, l2)
}
\arguments{
\item{a}{a p x k x n matrix of landmark coordinates}

\item{l1}{a numeric vector of landmark index (sequence number) to be changed}

\item{l2}{a numeric vector of landmark index (Sequence number) to change to}
}
\description{
Function to swap landmark configurations, or set as missing
}
\author{
Daniel Ehrlich
}
