% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_sym}
\alias{LMK_sym}
\title{Symmetrize}
\usage{
LMK_sym(A, LMpair)
}
\arguments{
\item{A}{A landmark configuration in the form p x k x n}

\item{LMpair}{A matrix listing paried landmarks, those not listed are assumed to be midline}
}
\value{
Returns a symmetrized landmark configuration with no missing data
}
\description{
Function to provide easy specification and handling of multistep symmetrization process
}
\details{
Under full sequence landmarks are first estimated using bilateral reflection. If missing data are still present, landmarks are estimated using TPS interpolation of n=3 nearest neighbors. Now complete LMK arrays are reflected a final time and averaged creating perfectly symmetrical configurations
}
