% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMK_Viz.R
\name{LMK_viz}
\alias{LMK_viz}
\title{Shape Change Visualizations}
\usage{
LMK_viz(dat, ...)
}
\arguments{
\item{dat}{Some form of landmark data of class array (p x k x n); gpagen (geomorph); PCA/CVA (morpho)}

\item{...}{Additional arguments to pass to subsequent functions based on input data}
}
\value{
will return a list of helpful information and some nice plots
}
\description{
Easily visualize shape change and create relative warps for LMK data
}
\details{
Attempt to combine all of these into a single function
}
