% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMK_plotting.R
\name{LMK_compare_two}
\alias{LMK_compare_two}
\title{Compare Two landmark arrays}
\usage{
LMK_compare_two(A1, A2, new = FALSE, main = NULL)
}
\arguments{
\item{A1}{A 3D array of landmark data; A reference}

\item{A2}{A second 3D array of landmark data to compare to the reference}

\item{new}{Should a new graphing window be opened (default is to overwrite the previous)}

\item{main}{A character string to display as the title. If main= NULL (default) function will attempt to deparse object names.}
}
\description{
Plot 2 landmark configurations in 3D space.
}
\details{
Presumably, this will be procrustes aligned coordinates, but any 2 p x k x n arrays will work
}
