% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_grp_apply}
\alias{LMK_grp_apply}
\title{Group Apply}
\usage{
LMK_grp_apply(d, f, grp)
}
\arguments{
\item{d}{a dataset}

\item{f}{a function}

\item{grp}{a grouping variable}
}
\description{
Apply a function (f) to a set of data (d) by a grouping variable (grp). Calculates values within group (eg mean), as well as for the entire ungrouped sample (eg, grand mean).
}
\examples{
data(iris)
LMK_grp_apply(d = iris[,1:4], f = mean, grp = iris$species) 

}
