% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_Screen}
\alias{LMK_Screen}
\title{Launch interactive screening plots}
\usage{
LMK_Screen()
}
\description{
GUI interface to choose DTA file, and various classifiers to conduct preliminary analysis.
}
