% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-support.R
\name{LMK_writeland_nts}
\alias{LMK_writeland_nts}
\title{Write DTA/NTS}
\usage{
LMK_writeland_nts(A, filepath)
}
\arguments{
\item{A}{a 3D array in the form of [p x k x n]}

\item{filepath}{Filename/Filepath, including extension: .dta, or .nts}
}
\description{
Write 3D array as a .dta/.nts file
}
\author{
Daniel Ehrlich
}
